Add Shortcut to parsers
Shortcut is working in progress.  Realized we need to also handle the if statements (refid alone is not enough - any combination of refid+if is a translatable string).  If statements have many possibilities (`if x.y.z.isMac() && isFirefox()`, etc) and can be generic so we don't know what'll be added later.  These are typically outlined in `org.rstudio.core.client.BrowseCap` and `org.rstudio.studio.client.application.Desktop`.  Need to consider this more.

Might be able to just concatenate refid+if then convert to only valid class name characters.  Is that workable?  It'll make for terrible class names in the interface/properties file, although maybe we could annotate them with better data in the comments